Commit 5
Each section now has four distinct colours corresponding to one of the four audio types. A fillColor variable has been added and assigned a hex value within each of the if/else if blocks, based on the dot's horizontal position. The fill(fillColor) function is added so that the colour is applied before its drawn on the canvas.