Update music_starter.js
This commit adds a dynamic multi-coloured effect to my visualiser, called "3d-starfield" inspired by a design I found on p5.sketch's community. The feature is implemented by dividing the screen into four quadrants, with each one containing a separate particle system that corresponds to one of the four audio variables: vocal, drum, bass, and other. The colour of the circles in each quadrant is mapped to the same colour scheme you used for your main visualizer, and their growth speed and size are directly controlled by the volume of their respective audio channel.